+++ CHANGELOG.md
@@ -11,3 +11,6 @@
 * Add Test API-Key options in settings for 
     * Klarna Pay Later
     * Afterpay
+* Add Test IP Address options in settings for 
+    * Afterpay
+    * Klarna Pay Later